% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviews_commands.R
\name{eviews_commands}
\alias{eviews_commands}
\title{Execute `EViews` commands from R}
\usage{
eviews_commands(commands="",wf="",page="")
}
\arguments{
\item{commands}{Object or a vector of character strings of `EViews` commands}

\item{wf}{Object or a character string representing the name of a workfile to be created}

\item{page}{Object or a character string representing the name of a workfile page to be created}
}
\value{
An EViews workfile
}
\description{
Use this function to execute `EViews` commands from R
}
\examples{
library(EviewsR)
\dontrun{
eviews_commands(c("wfcreate(wf=EViews_Workfile,page=Page) m 1990 2021","genr y=rnd","y.line"))
eviews_commands(c("genr y=rnd","y.line"),wf="EViews_Workfile")
# The first example creates an `EViews` workfile with monthly frequency from 1990 2021,
#
# then generate a series of random numbers named y, and finally plot a linegraph of the y series.

# The second example opens the `EViews` workfile and then generate a random series
# and plot its line graph.
}
}
\seealso{
eng_eviews, eviews_graph, eviews_import, eviews_object, eviews_pagesave, eviews_rwalk, eviews_wfcreate, eviews_wfsave, export, import_table, import
}
\keyword{documentation}
