% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviews_object.R
\name{eviews_object}
\alias{eviews_object}
\title{Create an `EViews` object on a workfile from R}
\usage{
eviews_wfcreate(wf_name="",page_name="",frequency="",start_date="",end_date="",path="",save=T)
}
\arguments{
\item{wf}{Object or a character string representing the name of an `EViews` workfile.}
}
\value{
An EViews workfile
}
\description{
Use this function to create an `EViews` object on a workfile from R
}
\examples{
library(EviewsR)
\dontrun{
eviews_wfcreate(wf_name="EVIEWSR_WORKFILE",page_name="EVIEWSR_PAGE",frequency="m",start_date="1990m1",end_date="2021m4",path="",save=T)
}
}
\seealso{
eng_eviews, eviews_commands, eviews_graph, eviews_import, eviews_object, eviews_pagesave, eviews_rwalk, eviews_wfcreate, eviews_wfsave, export, import_table, import
}
\keyword{documentation}
