% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviews_wfcreate.R
\name{eviews_wfcreate}
\alias{eviews_wfcreate}
\title{Create an `EViews` workfile from R}
\usage{
run_dynare(model,code,path)
}
\arguments{
\item{wf_name}{Object or a character string representing the name of a workfile to be created}

\item{page_name}{Object or a character string representing the name of a workfile page to be created}

\item{path}{Object or a character string representing the path to the folder for the \code{Dynare} file. The current working directory is the default `path`. Specify the `path` only if the `Dynare` files live in different path from the current working directory.}
}
\value{
Set of \code{Dynare} (open-source software for DSGE modelling) outputs
}
\description{
Use this function to create an `EViews` workfile from R
Use `run_dynare(model,code)`  if you want the `Dynare` files to live in the current working directory.
Use `run_dynare(model,code,path)` if you want the `Dynare` files to live in the path different from the current working directory.
}
\examples{
library(DynareR)
\dontrun{

}
}
\seealso{
eng_eviews
}
\keyword{documentation}
