% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviews_wfsave.R
\name{eviews_wfsave}
\alias{eviews_wfsave}
\title{Save an `EViews` workfile from R}
\usage{
eviews_wfsave(wf="",page="",options="",source_description="",
table_description="",keep_list="",drop_list="",keepmap_list="",dropmap_list="",
smpl_spec="")
}
\arguments{
\item{wf}{Object or a character string representing the name of an `EViews` workfile.}

\item{page}{Object or a character string representing the name of an `EViews` workfile page.}

\item{options}{Object or a character string of any of the acceptable `EViews` \code{pagesave} options, such as \code{noid}, \code{nomapval}, \code{nonames}.}

\item{source_description}{The path and name of the file to be saved.}

\item{table_description}{Further description of the \code{source_description} such as specifying the \code{range=arg}, \code{byrow}.}

\item{keep_list}{Optional. Specify the list of `EViews` object to be saved.}

\item{drop_list}{Optional. Specify the list of `EViews` object to be dropped.}

\item{keepmap_list}{Optional. Specify the list of patterns of `EViews` object to be saved.}

\item{dropmap_list}{Optional. Specify the list of patterns of `EViews` object to be dropped.}

\item{smpl_spec}{Optional. Specify the `EViews` sample string}
}
\value{
An EViews workfile.
}
\description{
Use this function to save an `EViews` workfile  from R
}
\examples{
library(EviewsR)
\dontrun{
 exec_commands(c("wfcreate(wf=Workfile,page=Page) m 1990 2022",
"genr y=rnd","genr x=rnd","save workfile","exit"))

eviews_wfsave(wf="workfile",source_description = "wfsave.csv",drop_list = "x")
}
}
\seealso{
eng_eviews, exec_commands, eviews_graph, eviews_import, create_object, eviews_pagesave, rwalk, eviews_wfcreate, export, import_table, import
}
\keyword{documentation}
