% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EviewsR-package.R
\docType{package}
\name{EviewsR-package}
\alias{EviewsR}
\alias{EviewsR-package}
\title{EviewsR: A Seamless Integration of 'EViews' and R}
\description{
It allows running 'EViews'(\url{https://eviews.com}) program from R, R Markdown and Quarto documents. 'EViews' (Econometric Views) is a statistical software for Econometric analysis. This package integrates 'EViews' and R and also serves as an 'EViews' Knit-Engine for 'knitr' package. Write all your 'EViews' commands in R, R Markdown or Quarto documents. For details, please consult our peer-review article Mati S., Civcir I. and Abba S.I (2023) \url{https://doi.org/10.32614/RJ-2023-045}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=EviewsR}
  \item Report bugs at \url{https://github.com/sagirumati/EviewsR/issues}
}


Other important functions: 
\code{\link{create_object}()},
\code{\link{eng_eviews}()},
\code{\link{eviews_graph}()},
\code{\link{eviews_import}()},
\code{\link{eviews_pagesave}()},
\code{\link{eviews_wfcreate}()},
\code{\link{eviews_wfsave}()},
\code{\link{exec_commands}()},
\code{\link{export_dataframe}()},
\code{\link{export}()},
\code{\link{import_equation}()},
\code{\link{import_graph}()},
\code{\link{import_kable}()},
\code{\link{import_series}()},
\code{\link{import_table}()},
\code{\link{import_workfile}()},
\code{\link{import}()},
\code{\link{rwalk}()},
\code{\link{set_eviews_path}()}
}
\author{
\strong{Maintainer}: Sagiru Mati \email{sagirumati@gmail.com} (\href{https://orcid.org/0000-0003-1413-3974}{ORCID})

}
\concept{important functions}
\keyword{documentation}
