% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_equation.R
\name{import_equation}
\alias{import_equation}
\title{Import `EViews` equation object as `kable`}
\usage{
import_equation(wf="",page="",equation_name="",equation_range="",format=kable_format(),
 digits = getOption("digits"), row.names = NA,col.names = NA, align,caption = NULL,
  label = NULL, format.args = list(),escape = FALSE, equation.attr = "", booktabs = TRUE,
   longequation = FALSE, valign = "t",position = "h", centering = TRUE,
   vline = getOption("knitr.equation.vline",if (booktabs) "" else "|"),
toprule = getOption("knitr.equation.toprule",
if (booktabs) "\\\\toprule" else "\\\\hline"),
bottomrule = getOption("knitr.equation.bottomrule",
if (booktabs) "\\\\bottomrule" else "\\\\hline"),
midrule = getOption("knitr.equation.midrule",
if (booktabs) "\\\\midrule" else "\\\\hline"),
linesep = if (booktabs) c("","", "", "", "\\\\addlinespace") else "\\\\hline",
 caption.short = "",equation.envir = if (!is.null(caption)) "equation",...)
}
\arguments{
\item{wf}{Object or a character string representing the name of a workfile to be created}

\item{page}{Object or a character string representing the name of a workfile page to be created}

\item{equation_range}{A vector of characters specifying the equation range of rows and columns}

\item{equation_name}{Name of an `EViews` equation object in an `EViews` workfile}
}
\value{
An EViews workfile
}
\description{
Use this function to import `EViews` equation object as `kable`
}
\examples{
library(EviewsR)
\dontrun{
demo(exec_commands)

import_equation(wf="EviewsR_exec_commands",page="page",equation_name="EviewsROLS",format="pandoc")
}
}
\seealso{
Other important functions: 
\code{\link{EviewsR}},
\code{\link{create_object}()},
\code{\link{eng_eviews}()},
\code{\link{eviews_graph}()},
\code{\link{eviews_import}()},
\code{\link{eviews_pagesave}()},
\code{\link{eviews_wfcreate}()},
\code{\link{eviews_wfsave}()},
\code{\link{exec_commands}()},
\code{\link{export_dataframe}()},
\code{\link{export}()},
\code{\link{import_dataframe}()},
\code{\link{import_graph}()},
\code{\link{import_kable}()},
\code{\link{import_series}()},
\code{\link{import_table}()},
\code{\link{import}()},
\code{\link{rwalk}()},
\code{\link{set_eviews_path}()}
}
\concept{important functions}
\keyword{documentation}
