% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_eviews_path.R
\name{set_eviews_path}
\alias{set_eviews_path}
\title{Set `EViews` path}
\usage{
set_eviews_path(eviews_system_path="eviews")
}
\arguments{
\item{eviews_system_path}{Path to the EViews executable}
}
\value{
Character
}
\description{
Use this function to set `EViews` path
}
\examples{
library(EviewsR)
\dontrun{
set_eviews_path('C:/Program Files (x86)/EViews 10/eviews10.exe')
}
}
\seealso{
eng_eviews, create_commands, eviews_graph, eviews_import, create_object, eviews_pagesave, rwalk, eviews_wfcreate, eviews_wfsave, export, import_table, import
}
\keyword{documentation}
