% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviews_commands.R
\name{eviews_commands}
\alias{eviews_commands}
\title{Execute `EViews` commands from R}
\usage{
eviews_commands(commands="",wf="",page="")
}
\arguments{
\item{commands}{Object or a vector of character strings of `EViews` commands}
}
\value{
An EViews workfile
}
\description{
Use this function to execute `EViews` commands from R
}
\examples{
library(EviewsR)
\dontrun{
eviews_commands(c("wfcreate(wf=page=eviews_workfile,page=eviews_page) m 1990 2021","genr y=rnd","y.line"))
eviews_commands(c("genr y=rnd","y.line"),wf="EviewsR")
}
}
\seealso{
eng_eviews, eviews_graph, eviews_import, eviews_object, eviews_pagesave, eviews_rwalk, eviews_wfcreate, eviews_wfsave, export, import_table, import
}
\keyword{documentation}
