% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviews_pagesave.R
\name{eviews_pagesave}
\alias{eviews_pagesave}
\title{Save an `EViews` workfile page from R}
\usage{
eviews_pagesave(wf="",page="",options="",source_description="",table_description="",keep_list="",drop_list="",keepmap_list="",
dropmap_list="",smpl_spec="")
}
\arguments{
\item{wf}{Object or a character string representing the name of an `EViews` workfile.}

\item{page}{Object or a character string representing the name of an `EViews` workfile page.}

\item{options}{Object or a character string of any of the acceptable `EViews` \code{pagesave} options, such as \code{noid}, \code{nomapval}, \code{nonames}.}

\item{source_description}{The path and name of the file to be saved.}

\item{table_description}{Further description of the \code{source_description} such as specifying the \code{range=arg}, \code{byrow}.}

\item{keep_list}{Optional. Specify the list of `EViews` object to be saved.}

\item{drop_list}{Optional. Specify the list of `EViews` object to be dropped.}

\item{keepmap_list}{Optional. Specify the list of patterns of `EViews` object to be saved.}

\item{dropmap_list}{Optional. Specify the list of patterns of `EViews` object to be dropped.}

\item{smpl_spec}{Optional. Specify the `EViews` sample string}
}
\value{
An EViews workfile.
}
\description{
Use this function to save an `EViews` workfile page from R
}
\examples{
library(EviewsR)
\dontrun{
eviews_pagesave(wf="workfile",source_description = "EviewsR.csv",drop_list = "y")
}
}
\seealso{
eng_eviews, eviews_commands, eviews_graph, eviews_import, eviews_object, eviews_pagesave, eviews_rwalk, eviews_wfcreate, eviews_wfsave, export, import_table, import
}
\keyword{documentation}
