% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_equation.R
\name{import_equation}
\alias{import_equation}
\title{Import `EViews` equation object as `kable`}
\usage{
import_equation(wf = "", page = "*", equation = "*")
}
\arguments{
\item{wf}{Object or a character string representing the name of a workfile to be created}

\item{page}{Object or a character string representing the name of a workfile page to be created}

\item{equation}{Name(s) of `EViews` equation object(s) in an `EViews` workfile}
}
\value{
An EViews workfile
}
\description{
Use this function to import `EViews` equation object as `kable`
}
\examples{
library(EviewsR)
\dontrun{
demo(exec_commands)

import_equation(wf="EviewsR_exec_commands",page="page",equation_name="EviewsROLS",format="pandoc")
}
}
\seealso{
Other important functions: 
\code{\link{EviewsR}},
\code{\link{create_object}()},
\code{\link{eng_eviews}()},
\code{\link{eviews_graph}()},
\code{\link{eviews_import}()},
\code{\link{eviews_pagesave}()},
\code{\link{eviews_wfcreate}()},
\code{\link{eviews_wfsave}()},
\code{\link{exec_commands}()},
\code{\link{export_dataframe}()},
\code{\link{export}()},
\code{\link{import_dataframe}()},
\code{\link{import_graph}()},
\code{\link{import_kable}()},
\code{\link{import_series}()},
\code{\link{import_table}()},
\code{\link{import_workfile}()},
\code{\link{import}()},
\code{\link{rwalk}()},
\code{\link{set_eviews_path}()}
}
\concept{important functions}
\keyword{documentation}
