% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_series.R
\name{import_series}
\alias{import_series}
\title{Import `EViews` series object as `kable`}
\usage{
import_series(wf="",page="",series_name="",series_range="",format=kable_format(),
 digits = getOption("digits"), row.names = NA,col.names = NA, align,caption = NULL,
  label = NULL, format.args = list(),escape = FALSE, series.attr = "", booktabs = TRUE,
   longseries = FALSE, valign = "t",position = "h", centering = TRUE,
   vline = getOption("knitr.series.vline",if (booktabs) "" else "|"),
toprule = getOption("knitr.series.toprule",
if (booktabs) "\\\\toprule" else "\\\\hline"),
bottomrule = getOption("knitr.series.bottomrule",
if (booktabs) "\\\\bottomrule" else "\\\\hline"),
midrule = getOption("knitr.series.midrule",
if (booktabs) "\\\\midrule" else "\\\\hline"),
linesep = if (booktabs) c("","", "", "", "\\\\addlinespace") else "\\\\hline",
 caption.short = "",series.envir = if (!is.null(caption)) "series",...)
}
\arguments{
\item{wf}{Object or a character string representing the name of a workfile to be created}

\item{page}{Object or a character string representing the name of a workfile page to be created}

\item{series_range}{A vector of characters specifying the series range of rows and columns}

\item{series_name}{Name of an `EViews` series object in an `EViews` workfile}
}
\value{
An EViews workfile
}
\description{
Use this function to import `EViews` series object as `kable`
}
\examples{
library(EviewsR)
\dontrun{
demo(exec_commands)

import_series(wf="EviewsR_exec_commands",page="page",series_name="EviewsROLS",format="pandoc")
}
}
\seealso{
Other important functions: 
\code{\link{EviewsR}},
\code{\link{create_object}()},
\code{\link{eng_eviews}()},
\code{\link{eviews_graph}()},
\code{\link{eviews_import}()},
\code{\link{eviews_pagesave}()},
\code{\link{eviews_wfcreate}()},
\code{\link{eviews_wfsave}()},
\code{\link{exec_commands}()},
\code{\link{export_dataframe}()},
\code{\link{export}()},
\code{\link{import_dataframe}()},
\code{\link{import_equation}()},
\code{\link{import_graph}()},
\code{\link{import_kable}()},
\code{\link{import_table}()},
\code{\link{import}()},
\code{\link{rwalk}()},
\code{\link{set_eviews_path}()}
}
\concept{important functions}
\keyword{documentation}
