% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_table.R
\name{import_table}
\alias{import_table}
\title{Import `EViews` table object as `kable`}
\usage{
import_table(wf="",page="",table_name="",table_range="",format=kable_format(),
 digits = getOption("digits"), row.names = NA,col.names = NA, align,caption = NULL,
  label = NULL, format.args = list(),escape = FALSE, table.attr = "", booktabs = TRUE,
   longtable = FALSE, valign = "t",position = "h", centering = TRUE,
   vline = getOption("knitr.table.vline",if (booktabs) "" else "|"),
toprule = getOption("knitr.table.toprule",
if (booktabs) "\\\\toprule" else "\\\\hline"),
bottomrule = getOption("knitr.table.bottomrule",
if (booktabs) "\\\\bottomrule" else "\\\\hline"),
midrule = getOption("knitr.table.midrule",
if (booktabs) "\\\\midrule" else "\\\\hline"),
linesep = if (booktabs) c("","", "", "", "\\\\addlinespace") else "\\\\hline",
 caption.short = "",table.envir = if (!is.null(caption)) "table",...)
}
\arguments{
\item{wf}{Object or a character string representing the name of a workfile to be created}

\item{page}{Object or a character string representing the name of a workfile page to be created}

\item{table_range}{A vector of characters specifying the table range of rows and columns}

\item{table_name}{Name of an `EViews` table object in an `EViews` workfile}
}
\value{
An EViews workfile
}
\description{
Use this function to import `EViews` table object as `kable`
}
\examples{
library(EviewsR)
\dontrun{
demo(exec_commands)

import_table(wf="EviewsR_exec_commands",page="page",table_name="EviewsROLS",format="pandoc")
}
}
\seealso{
Other important functions: 
\code{\link{EviewsR}},
\code{\link{create_object}()},
\code{\link{eng_eviews}()},
\code{\link{eviews_graph}()},
\code{\link{eviews_import}()},
\code{\link{eviews_pagesave}()},
\code{\link{eviews_wfcreate}()},
\code{\link{eviews_wfsave}()},
\code{\link{exec_commands}()},
\code{\link{export_dataframe}()},
\code{\link{export}()},
\code{\link{import_dataframe}()},
\code{\link{import_equation}()},
\code{\link{import_graph}()},
\code{\link{import_kable}()},
\code{\link{import_series}()},
\code{\link{import_workfile}()},
\code{\link{import}()},
\code{\link{rwalk}()},
\code{\link{set_eviews_path}()}
}
\concept{important functions}
\keyword{documentation}
