% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_eviews_path.R
\name{set_eviews_path}
\alias{set_eviews_path}
\title{Set `EViews` path}
\usage{
set_eviews_path(engine_path="eviews")
}
\arguments{
\item{engine_path}{Path to the EViews executable}
}
\value{
Character
}
\description{
Use this function to set `EViews` path
}
\examples{
library(EviewsR)
\dontrun{
set_eviews_path('C:/Program Files (x86)/EViews 10/eviews10.exe')
}
}
\seealso{
Other important functions: 
\code{\link{EviewsR}},
\code{\link{create_object}()},
\code{\link{eng_eviews}()},
\code{\link{eviews_graph}()},
\code{\link{eviews_import}()},
\code{\link{eviews_pagesave}()},
\code{\link{eviews_wfcreate}()},
\code{\link{eviews_wfsave}()},
\code{\link{exec_commands}()},
\code{\link{export_dataframe}()},
\code{\link{export}()},
\code{\link{import_graph}()},
\code{\link{import_kable}()},
\code{\link{import_series}()},
\code{\link{import_table}()},
\code{\link{import}()},
\code{\link{rwalk}()}
}
\concept{important functions}
\keyword{documentation}
